baokao_backend/pyproject.toml
[project]
name = "baokao-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.13"
Flask = "^3.1.0"
openai = "^1.64.0"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.2.0"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.1.1"
